# specs/001-rules-explorer/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Rules Explorer API
  description: API for D&D 5th Edition rules reference and search
  version: 1.0.0
servers:
  - url: /api
    description: Backend API
paths:
...
  /classes/{slug}:
    get:
      summary: Get class detail
      operationId: getClass
      tags:
        - Classes
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: Class detail with features
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /classes/{slug}/{subclassSlug}:
    get:
      summary: Get subclass detail
      operationId: getSubclass
      tags:
        - Classes
      parameters:
        - $ref: '#/components/parameters/Slug'
        - name: subclassSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Subclass detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubclassDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    Slug:
      name: slug
      in: path
      required: true
      schema:
        type: string
      description: URL-friendly identifier
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        code:
          type: string
    SourceCitation:
      type: object
      properties:
        document:
          type: string
          description: Source book name
        page:
          type: integer
          description: Page number
...
    ClassSummary:
      type: object
      required:
        - id
        - name
        - slug
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        hitDie:
          type: string
        primaryAbility:
          type: string
    ClassDetail:
      allOf:
        - $ref: '#/components/schemas/ClassSummary'
        - type: object
          properties:
            description:
              type: string
            savingThrows:
              type: array
              items:
                type: string
            armorProficiencies:
              type: array
              items:
                type: string
            weaponProficiencies:
              type: array
              items:
                type: string
            toolProficiencies:
              type: array
              items:
                type: string
            skillChoices:
              type: object
            features:
              type: array
              items:
                type: object
                properties:
                  level:
                    type: integer
                  name:
                    type: string
                  description:
                    type: string
            subclasses:
              type: array
              items:
                $ref: '#/components/schemas/SubclassSummary'
            source:
              $ref: '#/components/schemas/SourceCitation'
    SubclassSummary:
      type: object
      required:
        - id
        - name
        - slug
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
    SubclassDetail:
      allOf:
        - $ref: '#/components/schemas/SubclassSummary'
        - type: object
          properties:
            description:
              type: string
            features:
              type: array
              items:
                type: object
                properties:
                  level:
                    type: integer
                  name:
                    type: string
                  description:
                    type: string
            source:
              $ref: '#/components/schemas/SourceCitation'
            parentClass:
              $ref: '#/components/schemas/ClassSummary'
...